net(langage) : utilise intXX_t

+++ biblinternes/langage/erreur.hh
@@ -62,8 +62,8 @@ template <typename Flux>
 void imprime_caractere_vide(Flux &os, const int64_t nombre, const dls::vue_chaine &chaine)
 {
 	/* Le 'nombre' est en octet, il faut donc compter le nombre d'octets
-	 * de chaque point de code pour bien formater l'erreur. */
-	for (auto i = 0l; i < std::min(nombre, chaine.taille());) {
+     * de chaque point de code pour bien formater l'erreur. */
+    for (auto i = int64_t(0); i < std::min(nombre, chaine.taille());) {
 		if (chaine[i] == '\t') {
 			os << '\t';
 		}
